Rename Profiler to MetricViewer
